+++ src/main/java/fr/maxlego08/essentials/commands/commands/vault/CommandVaultAddSlot.java
@@ -3,6 +3,7 @@ package fr.maxlego08.essentials.commands.commands.vault;
 import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.CommandResultType;
 import fr.maxlego08.essentials.api.commands.Permission;
+import fr.maxlego08.essentials.api.messages.Message;
 import fr.maxlego08.essentials.vault.VaultModule;
 import fr.maxlego08.essentials.zutils.utils.commands.VCommand;
 import org.bukkit.OfflinePlayer;
@@ -14,7 +15,8 @@ public class CommandVaultAddSlot extends VCommand {
     public CommandVaultAddSlot(EssentialsPlugin plugin) {
         super(plugin);
         this.setModule(VaultModule.class);
-        this.setPermission(Permission.VAULT_ADD_SLOT);
+        this.setPermission(Permission.ESSENTIALS_VAULT_ADD_SLOT);
+        this.setDescription(Message.DESCRIPTION_VAULT_ADD_SLOT);
         this.addSubCommand("add");
         this.addRequireOfflinePlayerNameArg();
         this.addRequireArg("slot", (a, b) -> IntStream.range(1, 100).mapToObj(String::valueOf).toList());