+++ src/main/java/fr/maxlego08/essentials/vault/ZVault.java
@@ -5,6 +5,7 @@ import fr.maxlego08.essentials.api.vault.Vault;
 import fr.maxlego08.essentials.api.vault.VaultItem;
 import fr.maxlego08.essentials.zutils.utils.ZUtils;
 import fr.maxlego08.menu.zcore.utils.nms.ItemStackUtils;
+import org.bukkit.Material;
 import org.bukkit.inventory.ItemStack;
 
 import java.util.HashMap;
@@ -90,5 +91,16 @@ public class ZVault extends ZUtils implements Vault {
     @Override
     public int getNextSlot() {
         return IntStream.range(0, 45).filter(i -> !this.vaultItems.containsKey(i)).findFirst().orElse(-1);
+    }
+
+    @Override
+    public boolean hasFreeSlot() {
+        return getNextSlot() != -1;
+    }
+
+    @Override
+    public long getMaterialAmount(Material material) {
+        var optional = find(new ItemStack(material));
+        return optional.map(VaultItem::getQuantity).orElse(0L);
     }
 }